Restructure base calculation tests.
